Move `Edge` and `PertialEdgeInfo` to `network-protocol.rs`
