+++ chain/client/src/sync.rs
@@ -1290,7 +1290,7 @@ mod test {
 
     use super::*;
     use crate::test_utils::TestEnv;
-    use near_network::routing::PartialEdgeInfo;
+    use near_network::types::PartialEdgeInfo;
     use near_network::PeerInfo;
     use near_primitives::merkle::PartialMerkleTree;
     use near_primitives::types::EpochId;
