+++ chain/client/src/test_utils.rs
@@ -20,11 +20,10 @@ use near_chain::{
 };
 use near_chain_configs::ClientConfig;
 use near_crypto::{InMemorySigner, KeyType, PublicKey};
-use near_network::routing::PartialEdgeInfo;
 use near_network::test_utils::MockPeerManagerAdapter;
 use near_network::types::{
     FullPeerInfo, NetworkClientMessages, NetworkClientResponses, NetworkRecipient, NetworkRequests,
-    NetworkResponses, PeerManagerAdapter,
+    NetworkResponses, PartialEdgeInfo, PeerManagerAdapter,
 };
 use near_network::PeerManagerActor;
 use near_primitives::block::{ApprovalInner, Block, GenesisId};
