+++ chain/network/benches/routing_table_actor.rs
@@ -3,11 +3,11 @@ extern crate bencher;
 
 use bencher::{black_box, Bencher};
 use near_crypto::{KeyType, SecretKey, Signature};
-use near_network::routing::Edge;
 use std::collections::HashMap;
 use std::sync::Arc;
 
 use near_network::test_utils::random_peer_id;
+use near_network::types::Edge;
 use near_network::RoutingTableActor;
 use near_primitives::network::PeerId;
 use near_store::test_utils::create_test_store;
