+++ chain/network/src/routing/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod edge;
 pub(crate) mod edge_validator_actor;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub(crate) mod ibf;
@@ -6,16 +5,17 @@ pub(crate) mod ibf;
 pub(crate) mod ibf_peer_set;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub(crate) mod ibf_set;
+pub(crate) mod network_protocol;
 mod route_back_cache;
 pub(crate) mod routing;
 pub(crate) mod routing_table_actor;
 mod utils;
 
-pub use crate::routing::edge::{Edge, EdgeState, PartialEdgeInfo, SimpleEdge};
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub use crate::routing::ibf_peer_set::SlotMapId;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub use crate::routing::ibf_set::IbfSet;
+pub use crate::routing::network_protocol::{EdgeState, SimpleEdge};
 #[cfg(feature = "test_features")]
 pub use crate::routing::routing::GetRoutingTableResult;
 pub use crate::routing::routing::{
