+++ chain/network/src/tests/cache_edges.rs
@@ -1,4 +1,5 @@
-use crate::routing::{Edge, EdgeState, Prune, DELETE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME};
+use crate::network_protocol::Edge;
+use crate::routing::{EdgeState, Prune, DELETE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME};
 use crate::test_utils::random_peer_id;
 use crate::RoutingTableActor;
 use actix::System;
